+++ src/__tests__/GameSpec.ts
@@ -1655,7 +1655,7 @@ describe("test Game", () => {
 			game.pushScene(beforeScene);
 			game._flushPostTickTasks();
 
-			const currentScene = new Scene({ game, seeThrough: true });
+			const currentScene = new Scene({ game, seethrough: true });
 			const red = new FilledRect({
 				scene: currentScene,
 				cssColor: "red",
@@ -1666,7 +1666,7 @@ describe("test Game", () => {
 			game.pushScene(currentScene);
 			game._flushPostTickTasks();
 
-			// 現在のSceneのseeThroughがtrueの場合、1つ前のSceneから描画される
+			// 現在のSceneのseethroughがtrueの場合、1つ前のSceneから描画される
 			game.render();
 			expect(r.methodCallParamsHistory("fillRect").length).toBe(2);
 			expect(r.methodCallParamsHistory("fillRect")[0].cssColor).toBe("gray");